add filters for user vs service provided inputs, refactor db tests, add test for returning pipeline inputs to user
