+++ service/src/test/java/bio/terra/pipelines/service/PipelinesServiceTest.java
@@ -2,10 +2,8 @@ package bio.terra.pipelines.service;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
@@ -19,13 +17,11 @@ import bio.terra.pipelines.db.repositories.PipelineInputDefinitionsRepository;
 import bio.terra.pipelines.db.repositories.PipelinesRepository;
 import bio.terra.pipelines.testutils.BaseEmbeddedDbTest;
 import bio.terra.pipelines.testutils.TestUtils;
-import com.fasterxml.jackson.core.type.TypeReference;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.UUID;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
@@ -36,7 +32,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
 
 class PipelinesServiceTest extends BaseEmbeddedDbTest {
   @Autowired PipelinesService pipelinesService;
@@ -76,149 +71,19 @@ class PipelinesServiceTest extends BaseEmbeddedDbTest {
   }
 
   @Test
-  void allPipelineEnumsExist() {
-    // make sure all the pipelines in the enum exist in the table
-    for (PipelinesEnum p : PipelinesEnum.values()) {
-      assertTrue(pipelinesRepository.existsByName(p.getValue()));
-    }
-  }
+  void getPipelineInputDefinitions() {
+    PipelinesEnum imputationPipeline = PipelinesEnum.IMPUTATION_BEAGLE;
+    List<PipelineInputDefinition> allPipelineInputDefinitions =
+        pipelinesService.getAllPipelineInputDefinitions(imputationPipeline);
+    List<PipelineInputDefinition> userProvidedPipelineInputDefinitions =
+        pipelinesService.getUserProvidedInputDefinitions(imputationPipeline);
+    List<PipelineInputDefinition> serviceProvidedPipelineInputDefinitions =
+        pipelinesService.getServiceProvidedInputDefinitions(imputationPipeline);
 
-  @Test
-  void allPipelinesHaveDefinedInputs() {
-    // make sure all the pipelines in the enum have defined inputs
-    for (PipelinesEnum p : PipelinesEnum.values()) {
-      Pipeline pipeline = pipelinesRepository.findByName(p.getValue());
-      assertNotNull(pipeline.getPipelineInputDefinitions());
-    }
-  }
-
-  @Test
-  void allPipelineInputDefinitionsAreProperlyTyped() {
-    // make sure all pipeline inputs have defined types matching the enum
-    for (PipelineInputDefinition p : pipelineInputDefinitionsRepository.findAll()) {
-      assertDoesNotThrow(() -> PipelineInputTypesEnum.valueOf(p.getType()));
-    }
-  }
-
-  @Test
-  void allDefaultValuesForPipelineInputsAreCorrectType() {
-    // make sure all pipeline input definition default values pass type validation and are cast-able
-    for (PipelineInputDefinition p : pipelineInputDefinitionsRepository.findAll()) {
-      if (p.getDefaultValue() != null) {
-        PipelineInputTypesEnum inputType = PipelineInputTypesEnum.valueOf(p.getType());
-        assertNull(inputType.validate(p.getName(), p.getDefaultValue()));
-        assertNotNull(inputType.cast(p.getName(), p.getDefaultValue(), new TypeReference<>() {}));
-      }
-    }
-  }
-
-  @Test
-  void allOptionalAndServiceProvidedInputsHaveDefaultValues() {
-    // make sure all optional and service-provided inputs have default values
-    for (PipelineInputDefinition p : pipelineInputDefinitionsRepository.findAll()) {
-      if (!p.getIsRequired() || !p.getUserProvided()) {
-        assertNotNull(p.getDefaultValue());
-      }
-    }
-  }
-
-  @Test
-  void imputationPipelineHasCorrectInputs() {
-    // make sure the imputation pipeline has the correct inputs
-    Pipeline pipeline = pipelinesRepository.findByName(PipelinesEnum.IMPUTATION_BEAGLE.getValue());
-
-    List<PipelineInputDefinition> pipelineInputDefinitions = pipeline.getPipelineInputDefinitions();
-
-    // currently we have five inputs for the imputation pipeline
-    assertEquals(5, pipelineInputDefinitions.size());
-
-    // there should be one user-provided input and 4 service-provided inputs
     assertEquals(
-        1,
-        pipelineInputDefinitions.stream().filter(PipelineInputDefinition::getUserProvided).count());
-    assertEquals(
-        4,
-        pipelineInputDefinitions.stream()
-            .filter(Predicate.not(PipelineInputDefinition::getUserProvided))
-            .count());
-
-    // check user-provided inputs
-    assertTrue(
-        pipelineInputDefinitions.stream()
-            .filter(PipelineInputDefinition::getUserProvided)
-            .toList()
-            .stream()
-            .map(PipelineInputDefinition::getName)
-            .collect(Collectors.toSet())
-            .containsAll(Set.of("multi_sample_vcf")));
-
-    // check service-provided inputs
-    assertTrue(
-        pipelineInputDefinitions.stream()
-            .filter(Predicate.not(PipelineInputDefinition::getUserProvided))
-            .toList()
-            .stream()
-            .map(PipelineInputDefinition::getName)
-            .collect(Collectors.toSet())
-            .containsAll(
-                Set.of("contigs", "genetic_maps_path", "ref_dict", "reference_panel_path")));
-
-    // make sure the inputs are associated with the correct pipeline
-    assertEquals(
-        Set.of(pipeline.getId()),
-        pipelineInputDefinitions.stream()
-            .map(PipelineInputDefinition::getPipelineId)
-            .collect(Collectors.toSet()));
-  }
-
-  @Test
-  void addPipelineInput() {
-    Pipeline pipeline = pipelinesRepository.findByName(PipelinesEnum.IMPUTATION_BEAGLE.getValue());
-    List<PipelineInputDefinition> pipelineInputDefinitions = pipeline.getPipelineInputDefinitions();
-    assertEquals(5, pipelineInputDefinitions.size());
-
-    // add a pipeline input to the imputation pipeline
-    PipelineInputDefinition newInput = new PipelineInputDefinition();
-    newInput.setPipelineId(pipeline.getId());
-    newInput.setName("newInput");
-    newInput.setType(PipelineInputTypesEnum.INTEGER.toString());
-    newInput.setIsRequired(false);
-    newInput.setUserProvided(true);
-    newInput.setDefaultValue("42");
-
-    pipelineInputDefinitionsRepository.save(newInput);
-
-    pipeline = pipelinesRepository.findByName(PipelinesEnum.IMPUTATION_BEAGLE.getValue());
-    pipelineInputDefinitions = pipeline.getPipelineInputDefinitions();
-    assertEquals(6, pipelineInputDefinitions.size());
-
-    PipelineInputDefinition savedInput = pipelineInputDefinitions.get(5);
-    assertEquals("newInput", savedInput.getName());
-    assertEquals(PipelineInputTypesEnum.INTEGER.toString(), savedInput.getType());
-    assertFalse(savedInput.getIsRequired());
-    assertTrue(savedInput.getUserProvided());
-    assertEquals("42", savedInput.getDefaultValue());
-    assertEquals(pipeline.getId(), savedInput.getPipelineId());
-  }
-
-  @Test
-  void addDuplicatePipelineInputThrows() {
-    Pipeline pipeline = pipelinesRepository.findByName(PipelinesEnum.IMPUTATION_BEAGLE.getValue());
-    List<PipelineInputDefinition> pipelineInputDefinitions = pipeline.getPipelineInputDefinitions();
-    assertEquals(5, pipelineInputDefinitions.size());
-
-    // add a pipeline input that already exists
-    PipelineInputDefinition newInput = new PipelineInputDefinition();
-    newInput.setPipelineId(pipeline.getId());
-    newInput.setName("multi_sample_vcf");
-    newInput.setType(PipelineInputTypesEnum.INTEGER.toString());
-    newInput.setIsRequired(false);
-    newInput.setUserProvided(true);
-    newInput.setDefaultValue("42");
-
-    assertThrows(
-        DataIntegrityViolationException.class,
-        () -> pipelineInputDefinitionsRepository.save(newInput));
+        allPipelineInputDefinitions.size(),
+        userProvidedPipelineInputDefinitions.size()
+            + serviceProvidedPipelineInputDefinitions.size());
   }
 
   @Test
